+++ meta/CHANGELOG.md
@@ -2,9 +2,11 @@
 
 All notable changes to the Orbital Trading project will be documented in this file by the user, only. Gemini is never to modify or provide updates to this file, and must always treat this file as read-only. The user will always manually update this file.
 
+# [38.70] - 2026-6-22
+-  added sol station silhouette cinematic to mission 41
 
 # [38.69] - 2026-6-21
--  added kepler's eye cinematic
+-  added kepler's eye cinematic to mission 36
 -  mission bug fixes
 
 # [38.67] - 2026-6-20